==> src/car_price_prediction/pipeline/stage_02_prepare_base_model.py <==
from car_price_prediction.config.configuration import ConfigurationManager
from car_price_prediction.components.prepare_base_model import PrepareBaseModel
from car_price_prediction import logger

STAGE_NAME = "Prepare Base Model Stage"

class PrepareBaseModelTrainingPipeline:
    def __init__(self):
        pass

    def main(self):
        config = ConfigurationManager()
        prepare_base_model_config = config.get_prepare_base_model_config()
        prepare_base_model = PrepareBaseModel(config=prepare_base_model_config)
        
        logger.info("Creating base model")
        prepare_base_model.get_base_model()
        
        logger.info("Updating base model with training data")
        prepare_base_model.update_base_model()

if __name__ == '__main__':
    try:
        logger.info(f">>>>>> stage {STAGE_NAME} started <<<<<<")
        obj = PrepareBaseModelTrainingPipeline()
        obj.main()
        logger.info(f">>>>>> stage {STAGE_NAME} completed <<<<<<\n\nx==========x")
    except Exception as e:
        logger.exception(e)
        raise e
